11.3/ai-automation-hub/requirements.txt
fastapi==0.115.5
uvicorn==0.32.0
pydantic==2.9.2
python-dotenv==1.0.1
langchain==0.3.8
langchain-community==0.3.8
chromadb==0.5.18
sentence-transformers==3.2.1
httpx==0.27.2